refactor(tools): Moved Utility function
